requirements.txt
langchain-community==0.3.24
langchain-openai==0.3.21
streamlit==1.45.1
python-decouple==3.8